Split ButlerURI into separate files
Make new _butlerUri directory with files for each class providing an
implementation of a URI-handler
