+++ python/lsst/daf/butler/core/_butlerUri/__init__.py
@@ -1,0 +1,5 @@
+# Should only expose ButlerURI
+
+from ._butlerUri import ButlerURI
+__all__ = ("ButlerURI",)
+
