+++ python/lsst/daf/butler/core/_butlerUri/packageresource.py
@@ -1,0 +1,44 @@
+# This file is part of daf_butler.
+#
+# Developed for the LSST Data Management System.
+# This product includes software developed by the LSST Project
+# (http://www.lsst.org).
+# See the COPYRIGHT file at the top-level directory of this distribution
+# for details of code ownership.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import pkg_resources
+import logging
+
+from ._butlerUri import ButlerURI
+
+log = logging.getLogger(__name__)
+
+
+class ButlerPackageResourceURI(ButlerURI):
+    """URI referring to a Python package resource.
+
+    These URIs look like: ``resource://lsst.daf.butler/configs/file.yaml``
+    where the network location is the Python package and the path is the
+    resource name.
+    """
+
+    def exists(self) -> bool:
+        """Check that the python resource exists."""
+        return pkg_resources.resource_exists(self.netloc, self.relativeToPathRoot)
+
+    def read(self, size: int = -1) -> bytes:
+        with pkg_resources.resource_stream(self.netloc, self.relativeToPathRoot) as fh:
+            return fh.read(size)
